Use lower case values for state and attributes to match others from HA, Separate constants for dreamevacuum
